# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Implements the Mailio server based on https://mirs.mail.io/ specifications",
        "title": "Mailio Server API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/findaddress": {
            "get": {
                "description": "Returns a mailio address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Find user by base64 scrypt email address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64 formatted Scrypt of email address",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.OutputUserAddress"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/login": {
            "post": {
                "description": "Returns a JWS token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Login with username and password",
                "parameters": [
                    {
                        "description": "login input",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.InputLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.JwsToken"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing input parameters",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "401": {
                        "description": "Invalid signature",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "403": {
                        "description": "Failed to login (valid signature, no valid VC)",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "404": {
                        "description": "Failed to login (user not registered)",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/nonce": {
            "get": {
                "description": "Returns a nonce which client needs to sign with their private key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Login and Registration challenge nonce",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.NonceResponse"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
        "/api/v1/nonce/{id}": {
            "delete": {
                "description": "Deletes nonce if it exists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Deletes nonce if it exists",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nonce id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.NonceResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "description": "Returns a JWS token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "registration input",
                        "name": "registration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.InputRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.JwsToken"
                        }
                    },
                    "401": {
                        "description": "Invalid signature",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "404": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get logged in users basic information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Account"
                ],
                "summary": "Get logged inusers basic information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.OutputBasicUserInfo"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the HTTP status code",
                    "type": "integer"
                },
                "message": {
                    "description": "Message is the error message",
                    "type": "string"
                }
            }
        },
...
        "types.InputLogin": {
            "type": "object",
            "required": [
                "ed25519SigningPublicKeyBase64",
                "email",
                "mailioAddress",
                "nonce",
                "signatureBase64"
            ],
            "properties": {
                "ed25519SigningPublicKeyBase64": {
                    "description": "public key of the private key used to sign the nonce",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "mailioAddress": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "signatureBase64": {
                    "description": "signature of Nonce string",
                    "type": "string"
                }
            }
        },
        "types.InputRegister": {
            "type": "object",
            "required": [
                "databasePassword",
                "ed25519SigningPublicKeyBase64",
                "email",
                "mailioAddress",
                "nonce",
                "signatureBase64",
                "x25519PublicKeyBase64"
            ],
            "properties": {
                "databasePassword": {
                    "description": "this is a password for couchdbs private user database",
                    "type": "string"
                },
                "ed25519SigningPublicKeyBase64": {
                    "description": "public key of the private key used to sign the nonce",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "mailioAddress": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "signatureBase64": {
                    "description": "signature of Nonce string",
                    "type": "string"
                },
                "x25519PublicKeyBase64": {
                    "description": "public encryption key",
                    "type": "string"
                }
            }
        },
        "types.JwsToken": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "types.NonceResponse": {
            "type": "object",
            "properties": {
                "nonce": {
                    "type": "string"
                }
            }
        },
        "types.OutputBasicUserInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created": {
                    "type": "integer"
                },
                "totalDisk": {
                    "type": "integer"
                },
                "usedDisk": {
                    "type": "integer"
                }
            }
        },
        "types.OutputUserAddress": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}